requirements.txt
pypdf
langchain_openai
openai
openpyxl
reportlab
python-dotenv
langchain_community
langchain
requests
flask
datetime
openpyxl
gunicorn
pikepdf
pyodbc
bcrypt
Flask-Session

